DeepBlue Log Function
Created a log function that can be used globally across other log types by grabbing the LogType variable and LogFormat variable which a user will select. This function can be reused and will save many lines in the future.

Security log is built out, time to apply the changes to the other log types